refactor(perf): split src/acceptance/generator.ts — extract generator-helpers.ts
