+++ src/acceptance/generator.ts
@@ -10,6 +10,7 @@ import type { IAgentManager } from "../agents";
 import { getLogger } from "../logger";
 import type { UserStory } from "../prd/types";
 import { AcceptancePromptBuilder } from "../prompts/builders/acceptance-builder";
+import { extractTestCode, generateSkeletonTests } from "./generator-helpers";
 import {
   acceptanceTestFilename as defaultAcceptanceTestFilename,
   resolveAcceptanceTestFile as defaultResolveAcceptanceTestFile,
@@ -22,24 +23,7 @@ import type {
   RefinedCriterion,
 } from "./types";
 
-/**
- * Build the test framework import line for skeleton/fallback tests.
- *
- * Returns the appropriate import statement based on the detected framework,
- * or defaults to `bun:test` when none is specified.
- */
-function skeletonImportLine(testFramework?: string): string {
-  if (!testFramework) return `import { describe, test, expect } from "bun:test";`;
-  const fw = testFramework.toLowerCase();
-  if (fw === "jest" || fw === "@jest/globals") {
-    return `import { describe, test, expect } from "@jest/globals";`;
-  }
-  if (fw === "vitest") {
-    return `import { describe, test, expect } from "vitest";`;
-  }
-  // For other frameworks (e.g. "@testing-library/react"), keep bun:test base
-  return `import { describe, test, expect } from "bun:test";`;
-}
+export { extractTestCode, generateSkeletonTests } from "./generator-helpers";
 
 /**
  * Parse acceptance criteria from spec.md content.
@@ -514,191 +498,4 @@ export async function generateAcceptanceTests(
       criteria,
     };
   }
-}
-
-/**
- * Extract test code from agent output, supporting TypeScript, Go, Python, and Rust.
- *
- * Handles markdown code fences and extracts clean test code.
- *
- * @param output - Agent stdout
- * @returns Extracted test code
- */
-export function extractTestCode(output: string): string | null {
-  let code: string | undefined;
-
-  // Try to extract from markdown code fence (any language tag)
-  const fenceMatch = output.match(/```(?:\w+)?\s*([\s\S]*?)\s*```/);
-  if (fenceMatch) {
-    code = fenceMatch[1].trim();
-  }
-
-  // Go: package declaration followed by func Test
-  if (!code) {
-    const goMatch = output.match(/package\s+\w+[\s\S]*?func\s+Test\w+\s*\(/);
-    if (goMatch) {
-      const startIdx = output.indexOf(goMatch[0]);
-      code = output.slice(startIdx).trim();
-    }
-  }
-
-  // Python: def test_ function
-  if (!code) {
-    const pythonMatch = output.match(/(?:^|\n)((?:import\s+\w+[\s\S]*?)?def\s+test_\w+[\s\S]+)/);
-    if (pythonMatch) {
-      code = pythonMatch[1].trim();
-    }
-  }
-
-  // TypeScript: import statement
-  if (!code) {
-    const importMatch = output.match(/import\s+{[\s\S]+/);
-    if (importMatch) {
-      code = importMatch[0].trim();
-    }
-  }
-
-  // TypeScript: describe() block
-  if (!code) {
-    const describeMatch = output.match(/describe\s*\([\s\S]+/);
-    if (describeMatch) {
-      code = describeMatch[0].trim();
-    }
-  }
-
-  if (!code) return null;
-
-  // Validate: must contain at least one test-like keyword across all languages
-  const hasTestKeyword =
-    /\b(?:describe|test|it|expect)\s*\(/.test(code) || // TypeScript/JS
-    /func\s+Test\w+\s*\(/.test(code) || // Go
-    /def\s+test_\w+/.test(code) || // Python
-    /#\[test\]/.test(code); // Rust
-
-  if (!hasTestKeyword) {
-    return null;
-  }
-
-  return code;
-}
-
-/**
- * Generate skeleton acceptance tests with TODO placeholders.
- *
- * Used as fallback when LLM test generation fails.
- *
- * @param featureName - Feature name
- * @param criteria - Acceptance criteria to generate skeletons for
- * @returns TypeScript test code with TODO placeholders
- *
- * @example
- * ```ts
- * const skeleton = generateSkeletonTests("auth", [
- *   { id: "AC-1", text: "login succeeds", lineNumber: 5 },
- * ]);
- * // Generates test with TODO comment
- * ```
- */
-export function generateSkeletonTests(
-  featureName: string,
-  criteria: AcceptanceCriterion[],
-  testFramework?: string,
-  language?: string,
-): string {
-  const lang = language?.toLowerCase();
-
-  if (lang === "go") {
-    return generateGoSkeletonTests(featureName, criteria);
-  }
-
-  if (lang === "python") {
-    return generatePythonSkeletonTests(featureName, criteria);
-  }
-
-  if (lang === "rust") {
-    return generateRustSkeletonTests(featureName, criteria);
-  }
-
-  const tests = criteria
-    .map((ac) => {
-      return `  test("${ac.id}: ${ac.text}", async () => {
-    // TODO: Implement acceptance test for ${ac.id}
-    // ${ac.text}
-    expect(true).toBe(false); // Replace with actual test
-  });`;
-    })
-    .join("\n\n");
-
-  return `${skeletonImportLine(testFramework)}
-
-describe("${featureName} - Acceptance Tests", () => {
-${tests || "  // No acceptance criteria found"}
-});
-`;
-}
-
-function generateGoSkeletonTests(_featureName: string, criteria: AcceptanceCriterion[]): string {
-  const sanitize = (text: string) =>
-    text
-      .replace(/[^a-zA-Z0-9 ]/g, "")
-      .split(" ")
-      .map((w) => w.charAt(0).toUpperCase() + w.slice(1))
-      .join("");
-
-  const tests = criteria
-    .map((ac) => {
-      const funcName = `Test${sanitize(ac.text) || ac.id.replace("-", "")}`;
-      return `func ${funcName}(t *testing.T) {\n\t// TODO: ${ac.id}: ${ac.text}\n\tt.Fatal("not implemented")\n}`;
-    })
-    .join("\n\n");
-
-  return `package acceptance_test
-
-import "testing"
-
-${tests || "// No acceptance criteria found"}
-`;
-}
-
-function generatePythonSkeletonTests(_featureName: string, criteria: AcceptanceCriterion[]): string {
-  const sanitize = (text: string) =>
-    text
-      .toLowerCase()
-      .replace(/[^a-z0-9 ]/g, "")
-      .trim()
-      .replace(/\s+/g, "_");
-
-  const tests = criteria
-    .map((ac) => {
-      const funcName = `test_${sanitize(ac.text) || ac.id.toLowerCase().replace("-", "_")}`;
-      return `def ${funcName}():\n    # TODO: ${ac.id}: ${ac.text}\n    pytest.fail("not implemented")`;
-    })
-    .join("\n\n");
-
-  return `import pytest
-
-${tests || "# No acceptance criteria found"}
-`;
-}
-
-function generateRustSkeletonTests(_featureName: string, criteria: AcceptanceCriterion[]): string {
-  const sanitize = (text: string) =>
-    text
-      .toLowerCase()
-      .replace(/[^a-z0-9 ]/g, "")
-      .trim()
-      .replace(/\s+/g, "_");
-
-  const tests = criteria
-    .map((ac) => {
-      const funcName = sanitize(ac.text) || ac.id.toLowerCase().replace("-", "_");
-      return `    #[test]\n    fn ${funcName}() {\n        // TODO: ${ac.id}: ${ac.text}\n        panic!("not implemented");\n    }`;
-    })
-    .join("\n\n");
-
-  return `#[cfg(test)]
-mod tests {
-${tests || "    // No acceptance criteria found"}
-}
-`;
 }
